Submodule to handle exception have common functionalities. Wrapped in the 'utility' module. Original submodule 'utils.py' reallocated in 'utility module'. Rest of code refactored accordyingly.
