--- src/owimetadatabase_preprocessor/soil/utils/__init__.py
@@ -1,3 +1,0 @@
-"""A package to preprocess the data from owimetadatabase."""
-
-__version__ = "0.8.4"
